backend/requirements.txt
asgiref==3.9.1
boto3==1.40.11
botocore==1.40.11
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
django-ses==4.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
et_xmlfile==2.0.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
idna==3.10
jmespath==1.0.1
numpy==2.3.2
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.1
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
xlrd==2.0.2
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
geoip2==5.2.0
httpagentparser==1.9.5
django-redis==5.4.0
